refactor(modding): 加载与合并逻辑回归各类型自身，ModsManager 仅负责编排
